CODE FIXING

+++ src/dnsd.h
@@ -3,6 +3,7 @@
 
 #include <stdbool.h>
 #include <stdarg.h>
+#include <stddef.h>
 #define MAX_DN_LENGTH 255
 #define IP_LENGTH 4          /* For IPv4 */
 #define ANS_LENGTH 16        /* DNS answer */
@@ -82,6 +83,6 @@ void append_ln_to_log_file_dns_verbose(const char *msg, ...);
 static void vappend_ln_to_log_file_dns(const char *msg, va_list args);
 static void clear_log_file_dns();
 int get_machine_ip_dns(const char *iface, char *gateway_ip, size_t size);
-unsigned long remove_table(unsigned char *domain);
+unsigned int remove_table(unsigned char *domain);
 
 #endif /* DNSD_H */
